COMP: Fix compiler warning introduced by yesterday's commit.

+++ Utilities/vtkexodus2/excre.c
@@ -205,7 +205,7 @@ int ex_create_int (const char *path,
  */
 
 /* store Exodus API version # as an attribute */
-  vers = EX_API_VERS;
+  vers = (float) EX_API_VERS;
   if ((status=nc_put_att_float(exoid, NC_GLOBAL, ATT_API_VERSION,
 			       NC_FLOAT, 1, &vers)) != NC_NOERR)
   {
@@ -218,7 +218,7 @@ int ex_create_int (const char *path,
   }
 
 /* store Exodus file version # as an attribute */
-  vers = EX_VERS;
+  vers = (float) EX_VERS;
   if ((status=nc_put_att_float(exoid, NC_GLOBAL, ATT_VERSION, NC_FLOAT, 1, &vers)) != NC_NOERR)
   {
     exerrval = status;
